+++ console/src/main/java/org/example/ConsoleApplication.java
@@ -1,8 +1,6 @@
 
 package org.example;
 
-import org.example.bean.ModuleSrcConfigFile;
-import org.example.cache.FileCache;
 import org.example.exception.FileCacheException;
 import org.example.init.InitWorld;
 import org.example.init.ModuleSrcConfigFileInit;