+++ console/src/main/java/org/example/init/InitWorld.java
@@ -41,7 +41,6 @@ public class InitWorld {
                 return false;
             }
         }
-
         return true;
     }
 }